+++ apps/web/src/routes/docs.tsx
@@ -1,0 +1,9 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+import { DocsShell } from "../components/docs/DocsShell";
+
+export const Route = createFileRoute("/docs")({ component: DocsRoute });
+
+function DocsRoute() {
+  return <DocsShell />;
+}